+++ packages/interact/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @glissade/interact
 
+## 0.68.0-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.68.0-pre.0
+  - @glissade/player@0.68.0-pre.0
+  - @glissade/core@0.68.0-pre.0
+
 ## 0.67.0
 
 ### Patch Changes